LPD-68052 Move settings structure into the Settings > BasePage component to be able to create new structures like Settings > WizardPage
