+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/channels/pages/ChannelList.tsx
@@ -425,7 +425,6 @@ const ChannelList: React.FC<IChannelListProps> = ({
 
 	return (
 		<BasePage
-			groupId={groupId}
 			key='sitesListPage'
 			pageDescription={
 				<>
