+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/definitions/event-attributes/pages/AttributeView.tsx
@@ -76,7 +76,6 @@ const AttributeView: React.FC<IAttributeViewProps> = ({
 						getEventAttributes({groupId}),
 						{active: true, label: name}
 					]}
-					groupId={groupId}
 					pageActions={viewAttributePageActions}
 					pageDescription={
 						<>
