+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/definitions/events/pages/View.tsx
@@ -60,7 +60,6 @@ const View: React.FC<IViewProps> = ({close, eventId, groupId, open}) => {
 						getEvents({groupId}),
 						{active: true, label: displayName || name}
 					]}
-					groupId={groupId}
 					pageActions={viewEventPageActions}
 					pageDescription={
 						description || Liferay.Language.get('no-description')
