+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/UsageOverview.tsx
@@ -65,7 +65,7 @@ const getAlertStatusCode = currentPlan => {
 	return null;
 };
 
-export const UsageOverview = ({groupId, project}) => {
+export const UsageOverview = ({project}) => {
 	const [showAlert, setShowAlert] = useState(true);
 	const currentUser = useCurrentUser();
 	const admin = currentUser.isAdmin();
@@ -96,7 +96,6 @@ export const UsageOverview = ({groupId, project}) => {
 
 	return (
 		<BasePage
-			groupId={groupId}
 			key='UsageOverview'
 			pageActions={pageActions}
 			pageDescription={Liferay.Language.get(
