+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/UsageOverviewSaaS.tsx
@@ -14,11 +14,7 @@ export type Resource = {
 	measurement: string;
 };
 
-interface IUsageOverviewSaaS {
-	groupId: string;
-}
-
-export const UsageOverviewSaaS: React.FC<IUsageOverviewSaaS> = ({groupId}) => {
+export const UsageOverviewSaaS = () => {
 	const currentUser = useCurrentUser();
 
 	let pageActions = [];
@@ -60,7 +56,6 @@ export const UsageOverviewSaaS: React.FC<IUsageOverviewSaaS> = ({groupId}) => {
 
 	return (
 		<BasePage
-			groupId={groupId}
 			key='UsageOverview'
 			pageActions={pageActions}
 			pageDescription={Liferay.Language.get(
