+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/pages/user/index.tsx
@@ -65,7 +65,6 @@ export const User = ({className}) => {
 	return (
 		<BasePage
 			className={getCN('user-list-page-root', className)}
-			groupId={groupId}
 			key='userListPage'
 			pageDescription={Liferay.Language.get(
 				'invite-new-users-to-analytics-cloud-and-or-configure-existing-users'
