+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/recommendations/pages/CreateItemSimilarity.tsx
@@ -25,7 +25,6 @@ const CreateItemSimilarity: React.FC<ICreateItemSimilarityProps> = ({
 					label: Liferay.Language.get('new-model')
 				}
 			]}
-			groupId={groupId}
 			pageDescription={Liferay.Language.get(
 				'item-similarity-model-uses-items-and-iteractions-for-training'
 			)}
