+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/recommendations/pages/Recommendations.tsx
@@ -18,7 +18,6 @@ const Recommendations: React.FC<IRecommendationsProps> = ({
 
 	return (
 		<BasePage
-			groupId={groupId}
 			pageDescription={Liferay.Language.get(
 				'create-and-train-machine-learning-models-to-use-in-your-recommendations'
 			)}
