+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/settings/recommendations/pages/View.tsx
@@ -88,7 +88,6 @@ const View: React.FC<IViewProps> = ({addAlert, close, history, job, open}) => {
 							label: name
 						}
 					]}
-					groupId={groupId}
 					pageActions={
 						currentUser.isAdmin()
 							? [
